refactor: remove input_mutation and move InputMutationExtension to a new module
