+++ RELEASE.md
@@ -1,12 +1,16 @@
 Release type: minor
 
-This release adds `strawberry.input_mutation` as a way to make it easier to create
-mutations that receive a single input type called `input`, while still being able
-to define the arguments of that input in the resolver itself.
+This release adds a new field extension called `InputMutationExtension`, which makes
+it easier to create mutations that receive a single input type called `input`,
+while still being able to define the arguments of that input on the resolver itself.
 
 The following example:
 
 ```python
+import strawberry
+from strawberry.field_extensions import InputMutationExtension
+
+
 @strawberry.type
 class Fruit:
     id: strawberry.ID
@@ -16,12 +20,15 @@ class Fruit:
 
 @strawberry.type
 class Mutation:
-    @strawberry.input_mutation
+    @strawberry.mutation(extensions=[InputMutationExtension()])
     def update_fruit_weight(
         self,
         info: Info,
         id: strawberry.ID,
-        weight: float,
+        weight: Annotated[
+            float,
+            strawberry.argument(description="The fruit's new weight in grams"),
+        ],
     ) -> Fruit:
         fruit = ...  # retrieve the fruit with the given ID
         fruit.weight = weight
@@ -34,6 +41,10 @@ Would generate a schema like this:
 ```graphql
 input UpdateFruitInput {
   id: ID!
+
+  """
+  The fruit's new weight in grams
+  """
   weight: Float!
 }
 
@@ -48,5 +59,6 @@ type Mutation {
 }
 ```
 
-That pattern makes it easier to include/remove arguments without breaking the
-whole API, which could happen when using positional arguments.
+This pattern makes it easier to include/remove arguments without breaking the
+whole API, which could happen when using positional arguments. Additionally,
+fields in an input type can be marked as deprecated, while field can't be deprecated.
