+++ strawberry/field_extensions/__init__.py
@@ -1,0 +1,5 @@
+from .input_mutation import InputMutationExtension
+
+__all__ = [
+    "InputMutationExtension",
+]
